Separate performance contracts and adapters
