+++ core/common/src/main/kotlin/sk/styk/martin/apkanalyzer/core/common/performance/PerformanceTrace.kt
@@ -1,0 +1,7 @@
+package sk.styk.martin.apkanalyzer.core.common.performance
+
+interface PerformanceTrace : AutoCloseable {
+    fun putMetric(name: String, value: Long)
+
+    fun putAttribute(name: String, value: String)
+}
