+++ core/common/src/main/kotlin/sk/styk/martin/apkanalyzer/core/common/performance/PerformanceTracker.kt
@@ -3,9 +3,3 @@ package sk.styk.martin.apkanalyzer.core.common.performance
 interface PerformanceTracker {
     fun startTrace(name: String): PerformanceTrace
 }
-
-interface PerformanceTrace : AutoCloseable {
-    fun putMetric(name: String, value: Long)
-
-    fun putAttribute(name: String, value: String)
-}
